Setup testing and start mocking node-fetch
