+++ __mocks__/node-fetch.js
@@ -1,0 +1,3 @@
+const fetch = (url);
+
+module.exports = fetch;
