+++ package.json
@@ -7,7 +7,7 @@
     "start": "node server/index.js",
     "start-dev": "nodemon server/index.js --inspect",
     "lint": "eslint ./ --ext .jsx --ext .js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest test"
   },
   "repository": {
     "type": "git",
@@ -39,9 +39,22 @@
     "eslint-plugin-import": "^2.10.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.7.0",
+    "jest": "^22.4.3",
     "nodemon": "^1.17.3",
     "regenerator-runtime": "^0.11.1",
+    "supertest": "^3.0.0",
     "webpack": "^4.5.0",
     "webpack-cli": "^2.0.14"
+  },
+  "jest": {
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "client/src/**/*.jsx",
+      "db/*.js",
+      "server/*.js",
+      "!db/seed.js",
+      "!client/src/index.jsx",
+      "!server/index.js"
+    ]
   }
 }
